% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runexample.R
\name{aboutApp}
\alias{aboutApp}
\title{aboutApp : get more information about an App !}
\usage{
aboutApp(appname)
}
\arguments{
\item{appname}{string of which app you want to get info about
This function opens the description of any given app}
}
\description{
aboutApp : get more information about an App !
}
