% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runexample.R
\name{launchApp}
\alias{launchApp}
\title{launchApp: Launch an App!}
\usage{
launchApp(appname)
}
\arguments{
\item{appname}{string of which app you want to run
This function runs the shiny app for a given `appname`. run without an argument \code{launchApp()} to see a list of available apps.}
}
\description{
launchApp: Launch an App!
}
