% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runexample.R
\name{runExample}
\alias{runExample}
\title{runExample: Run Examples!}
\usage{
runExample(example)
}
\description{
This function runs the shiny app for a given `example`. run without an argument \code{runExample()} to see a list of available apps.
}
