% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runexample.R
\name{runTutorial}
\alias{runTutorial}
\title{runTutorial: Run a Tutorial!}
\usage{
runTutorial(tutoname)
}
\arguments{
\item{tutoname}{string of which tutorial you want to run
This function runs a given `tutoname` for you. run without an argument \code{runTutorial()} to see a list of available tutorials.}
}
\description{
runTutorial: Run a Tutorial!
}
