% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScPoEconometrics-package.R
\docType{package}
\name{ScPoEconometrics}
\alias{ScPoEconometrics}
\alias{ScPoEconometrics-package}
\title{ScPoEconometrics: Teaching Econometrics with R at SciencesPo}
\description{
The is the R package for the 2nd year UG
course in econometrics at SciencesPo.
}
\keyword{internal}
